Move Pic component into common and make index export it as our entry point
